+++ ios/Classes/Src/ObjectConverter.swift
@@ -279,6 +279,12 @@ extension PanoWBDocContents {
         if let urls = map["urls"] as? [String] {
             self.urls = urls
         }
+        if let thumbUrls = map["thumbUrls"] as? [String] {
+            self.thumbUrls = thumbUrls
+        }
+        if let docId = map["docId"] as? String {
+            self.docId = docId
+        }
     }
 }
 
@@ -307,6 +313,21 @@ extension PanoWBVisionConfig {
         }
         if let limited = map["limited"] as? Bool {
             self.limited = limited
+        }
+    }
+}
+
+extension PanoWBDocExtHtml {
+    convenience init(map: Dictionary<String, Any>) {
+        self.init()
+        if let name = map["name"] as? String {
+            self.name = name
+        }
+        if let url = map["url"] as? String {
+            self.url = url
+        }
+        if let thumbUrls = map["thumbUrls"] as? [String] {
+            self.thumbUrls = thumbUrls
         }
     }
 }